Start adding actual data
Bunch of stuff going on here. First, this introduces some standardized
structure for defining models (complete with schemas! and inheritance!
and optionally instance methods!).

Next, it introduces the basic types for hunts (one goal for JR is to
have multiple distinct hunts in the same database) and puzzles (which
are represented as a tree, with the hunt object as the root).

Finally, a simple starting point for rendering the hunt puzzle tree.