Add a modal for creating a new hunt
By reusing the same template for both the new hunt and list hunts
route, and including a modal in the template, we can use a separate
route for the new hunt page but have it work as a modal.

I think this is a style we'll want for a bunch of editing constructs,
though it wasn't totally straightforward. Notably, I couldn't use
data-dismiss, which the Bootstrap docs recommend, since that would
allow Bootstrap to get out of sync with Meteor state.