Add "Ansible" for talking instantly across the ether
Logging. Logging is pretty cool. We should make it easy to log. And
since most of the semantic awareness is on the client, we should make
it easy for the client to tell the server what it's doing.

Use `Ansible.log` the same way you'd use `console.log`, and it'll
magically get sent via quantum entanglement (i.e. Meteor method calls)
to the server and printed out as well.