Add a login page
It works by reactively swapping out the standard ReactRouter.Router
for a Login component when the user is not logged in. We'll need
something more complicated once we start working on the onboarding
flow (likely another router tree?), but this at least lets you sign in
at all.

I'm _super_ unconvinced that this is the right way to approach this,
but it does seem to work, and has the notable advantage that you don't
have to worry about missing edge transitions when logging in and
logging out.