WIP on components
I added a dependency on Radium so I can do a bunch of stuff with inline styles
more easily.  Inline styles can actually be typechecked and validated, as well
as easily reused with the same tooling that we use for our JS modules.

If this makes people cringe too much, I guess I can deal with writing separate
CSS, it's just nicer to have all the styles right there with the component
that's being styled.