Add page for sending an invite email
Still an open question what happens if you try to send an invite to an
account that already exists. In the old site, that triggers a password
reset.

I suspect the right logic is: if the user hasn't clicked through their
confirmation email, send another one. Otherwise, return an error
message to that effect and suggest the forgot password flow (which is
separate).