Flesh out the account enrollment flow
First, allow re-sending the enrollment email if necessary, as long as
the account hasn't yet been claimed. Add routes for the enrollment URL
and reset URL, and register them with the accounts email
templates. Finally, fix the onLogin logging, which was pretty finely
tuned to actual logins and not other logging-in events (like password
resets).