Get ELB + LetsEncrypt working
This relies on having the LetsEncrypt client throw challenges into an
S3 bucket, which _all_ frontends will redirect to. We manage the
LetsEncrypt issuance ourselves rather than using the nginx+letsencrypt
docker container.

Still need to deal with DNS and cert renewal.

As a note, it's currently pointed at the LetsEncrypt staging server,
because I didn't realize the rate limits were so tough, and I wanted
to keep testing.