Initial implementation of ShareJS-synchronized spreadsheet
This supports editing cells, but not a lot more. Things like inserting
and deleting rows and columns still doesn't work (and will cause
Handsontable to get out of sync with ShareJS's state).

It's a bit of an open question for me how to handle inserting columns,
since Handsontable doesn't give us a hook to interrupt that like it
does with deletions and modifications, but maybe we can come up with
something.