Add support for adding/removing rows
Handsontable.Core.alter is how UI code triggers changes to the table
topography, so intercept that and submit the changes to ShareJS
instead.

This doesn't implement add/remove columns. Given the current data
representation, in order to do it we'd have to submit a separate
operation to add the column to each row, and that won't compose
correctly with other operations. I think we may have to define a
custom OT to represent that (though it should be fairly
straightforward).