Use a custom operational transform for grids
The JSON OT type isn't sufficient for us to perform column
insertions/deletions in a concurrency-friendly way, so we need a
custom OT type. This defines column insertion/deletion as native
operations, as well as row insertion/deletion, and cell set.

I haven't implemented the transformation for column/row delete yet,
which means that performing these operations concurrently will cause
problems. (Performing them non-concurrently is fine, since
transformations are only required to resolve concurrent edits)

The actual canonical representation of an operation is a list of
individual component operations, because otherwise it's impossible to
express composition or inversion. The json0 OT type already has some
code for wiring this type of OT together, so I just pull that in.