Enable formulas
Because we prevent Handsontable from making changes directly to the
underlying data structure, the normal hooks for inserting and removing
a column don't fire (since, as far as it can tell, they happen as
completely opaque changes to the underlying data structure).

To fix that, we manually trigger the hooks after applying ops, which
makes the formula hooks work. There's a separate issue
(handsontable/handsontable-ruleJS#19) that removing columns and rows
doens't do anything, but we can deal with that later.

I also suspect that the beforeAutofillInsidePopulate hook isn't
working correctly, and so formulas aren't getting updated when you
fill them.