Replace AccountsTemplates forms with our own
We need to be able to collect additional information in the onboarding flow.
While AccountsTemplates has API for letting you add input fields to a signup form,
none of our workflows actually use the appropriate form.

So instead we produce similar DOM (but with our needed additions) ourselves,
and make the appropriate Meteor Accounts calls and our own Meteor method calls
directly.  This allows us to collect more information in the enrollment flow,
which lets us prepopulate profiles with useful information.

While this change means we're not using AccountsTemplates directly, we're still
using the CSS in our new AccountForm, so I've left the package in.  A refactoring
would let us drop AccountsTemplates, which would save a minified ~150kb on pageload,
since we'd also drop softwarerero:accounts-t9n which has 117kb of localizations.