Allow numbers in answers
I think I remember seeing puzzles in past hunts that used numbers in the
answer, so we should probably allow those, even if it's bad hunt form to
do so.  Worst case, if the actual hunt site won't allow them, we can reject
answers with numbers in them from the operator queue.