Create a Google Spreadsheet for each puzzle
This is the big one, and it's got a fair number of moving
parts. First, we thread the credentials that we collect from /setup
through the Google Drive REST API, making sure to update our copy of
the credentials should they need refreshing.

Then we add a method which creates a document for a puzzle if one does
not already exist.

Finally, we load that document on the puzzle page and thread it
through, until we finally render an iframe with the document.