Fix exception when navigating from one puzzle page to another
We were guarding the creation of the profiles object with whether or not chat
had loaded, but this is too aggressive an optimization: rendering past guesses
also requires the profiles to exist.

So, we should make the "main puzzles view" also depend on the profile handle
being ready, and then we should release the profile dependency on chat.

This changeset now avoids building the lists for puzzles/tags/guesses/documents
if they're not going to be used because we're still showing a loading message,
which should improve clientside performance by reducing rerenders as data
streams in over DDP substantially, especially for larger hunts.