Make the Slack and Guess notifications hideable client-side
The Slack notification, in particular, is obnoxious: I cannot log out or even
navigate to the settings page myself because the notification is covering the
menu dropdown.

Guesses are slightly better, in that I can hide them all by disabling operator
mode, but they have a similar problem.

Notifications should basically *always* be dismissable, or they will cover
important things on the page and users won't be able to do the thing they care
about, because an undismissable notification is in the way.