Rework how subscriber counters work to make fetching them more efficient
When you increment a counter, you include a "context" object that gets
stored with the counter record. When you fetch counters, you provide a
query on the context object so that bulk fetches become more
efficient.

There's probably still going to be a reasonable cost to aggregating
the counters in the subCounter.fetch implementation, but at our scale
it shouldn't be too big of a deal.