Drop Kadira integration
Kadira's instrumentation seems to be rather unhappy when we do certain
invocations of findOne(), which makes Kadira crash (and fails our method call)
when things work fine if I disable the Kadira hooks.

Presumably we'll want to reenable the integration, but this is a pain to deal
with at the moment, so I'm disabling it for now.