Rewrite notifications center sans the messenger dependency
Messenger is almost completely unused in the current implementation - we use it
exclusively for some CSS.  It's also the only remaining dependency which we
pull from bower, which I'd like to move away from.

I wrote my own CSS to imitate Messenger's look and feel, and applied it in
NotificationCenter, then everything that has to do with bower.

This patch also switches the sort order of messages to stack newest below
oldest, so that existing messages don't move when a new guess arrives in the
queue.  Since the guess queue is a queue, not a stack, this seemed reasonable.