Move from <BS.Input> to <BS.FormControl>
react-bootstrap deprecated <Input> in favor of a lower-level API which more
closely models the underlying DOM:

  https://github.com/react-bootstrap/react-bootstrap/issues/1749

This patch moves all of our inputs to use the lower-level API, so when we
upgrade to 0.30 (and <BS.Input> is removed entirely) things don't break.

While we're moving inputs about, reduce the usage of refs.  Event handlers
get the new value of input fields in the event object, so a bunch of things
we did with refs can be simplified.