Don't setup our own listener with nginx-https
Now that we're no longer doing PROXY protocol, we can just use the
standard listener config for nginx-proxy.

Also a few other cleanups. Now that we're no longer terminating TLS on
the app server, we need to set the HSTS header ourselves. And ALB sets
X-Forwarded-Proto, so we don't need to override it.