Log failed logins directly, not through ansible
The onLoginFailure hook gets called in a client-initiated
context (i.e. `this.connection` is set), which means that Ansible
rejects it (since the user isn't logged in, obviously - the login just
failed).

We could solve this by making the Ansible interface a bit more flexible,
but realistically it's not worth the effort, so just splat things out in
place.