Add new addToHunt method for adding another user
When you call addToHunt, you specify an email address of another user
(which may not yet exist). We create the user if it doesn't already
exist, add the user to the hunt either way, and send them an
email. This also significantly refactors our signup email to indicate
the modifications we've already done, and adds a new email template
for adding an existing user to a new hunt.