Try to make activity tracking more efficient
There were a couple things we were doing that were terrible:

- Every time we ran the periodic cleanup, we'd write to every
  Subscribers record created on that server.

- We ran the periodic function every 2.5 _milli_seconds

Even if the periodic function was running every 2.5 seconds as
intended, that would probably be much too frequent - the mechanisms in
the periodic code only matter if a server actually exits. Normal
operations will get cleaned up by the onStop handler.

Therefore, we make the timeouts and periodic frequencies much less
agressive (and correct the units). And to reduce the write fanout,
track whether a given server is still alive and active in a separate,
new Servers collection.