Make puzzles editable
This sets up an edit modal for puzzles in puzzle lists and on the
puzzle page. It also consolidates the arguments we pass to the
createPuzzle and updatePuzzle server methods to require a little less
moving state around.

Finally, it updates the name of the associated Google document (which,
for the record, is a horrendously poorly documented thing). I don't
think there's any other existing state we need to update in this case.