Get rid of hunt landing page
Instead, redirect people directly to the puzzle list. Now that this
always includes a link to announcements and the guess queue in the top
right, it has all the functionality the hunt landing page ever did.

(I think I didn't realize this was even there because prior to
0e144376, those links were replaced on the operator view by the Add
puzzle button. Now they always display, and the add puzzle button goes
in a separate slot.)