Set a timer to refresh gdrive oauth credentials when they expire
If we're actively making API calls, the googleapis library will
automatically refresh the oauth credentials as needed. If Jolly Roger
is idle for an extended period, though, this might not
happen. Whenever we get new credentials, look at when they expire, set
a timer for some brief period after that, and manually refresh.

The reactivity here should keep this from firing spuriously, and the
jitter should keep multiple servers/processes from hammering Google
simultaneously (which probably works, but is avoidable).