Allow users to link a Google account to their profile
This goes through a Google OAuth identity flow, so we get a validated
email address, but throws away the resulting OAuth credentials.

There are two main ways this is useful. First, we can use this with
the Google Drive API to explicitly share documents with the users
viewing the document, which should get rid of the "Anonymous
Animals". Second, we can share the list with
practice.deathandmayhem.com for its auth.