Start to sketch out a user existence API for practice
We define "active" users (or, really, active team mebers) to be users
who have looked at Jolly Roger in the last year. This should account
for people who logged in at last year's hunt.

We could probably shorten it slightly (11 months and assume that we'll
start the signup process early enough?) but that seems a little user
hostile, and this has the property that it still culls people who
don't hunt with us.