Rework how Google OAuth works again
From reading the docs, it seems like the expectation is that refresh
tokens are the only credential that's stored, and the client is
expected to more or less always have to fetch a new access token at
startup.

So do that, and don't bother persisting access tokens. Schedule a
refresh roughly 1/4 of the way through an access token's
lifetime (roughly 15 minutes) to fetch a new one. Each server process
has its own in-memory access token.