Rework fixtures to just be a meteor method
When we first wrote the fixtures, we didn't have a fully-operational backend
and data structures, but wanted to experiment with UI.  Now that we have a
functional backend, we don't need to be shipping an entire hunt's worth of
fixture data to every client in the JS bundle.

This reduces the client bundle size by ~50k.