Make Puzzle pure again
This patch renames a bunch of instances of `tags` to `allTags`, because that's
what they are (all tags associated with this hunt), and we seem to have been
confused and fetched all the tags again despite already having a copy of all
the tags in props.

This patch also pushes canEdit down through the puzzle component hierarchy.

Together, this change shaves about 10ms off the PuzzleListPage render time
for the 2015 hunt dataset.