Move a bunch of styles from inline styles to CSS
My hope here is that by reducing the number of elements that React
needs to set inline styles on, we can reduce the amount of time spent
rendering the puzzle list page.

If that doesn't pan out in practice, my other hope is that our DOM stops
looking like a crime against humanity.

I left a few inline styles in places where:

* I wasn't sure what the right thing to do is (ReactSelect2)
* I expect to want to make certain aspects of appearance user-controlled (split
  width on PuzzlePage)
* We weren't using the code anyway (LabelledRadioGroup)