Consolidate ChatMessages back into using Schema.Base
We broke it out of Schema.Base in anticipation of bidirectional chat
syncing from Slack. However, we ended up not doing that and just using
the firehose, and we don't think we're likely to change our minde, so
switch back to using the standard base model.

This also eliminates some auxiliary schemas/models (ChatMetadata and
SlackMessageInfo) that were similarly tied to the idea of
bidirectional Slack syncing.

Because models are now expected to have a createdAt and createdBy, we
backfill those. Because they can be destroyed, we need to backfill
`deleted` to false. In order to satisfy our standard publish security
model, we backfill a `hunt` property.

And since we're already rewriting everything, change the `puzzleId`
prop to `puzzle`. The old name was a bit clearer, but inconsistent
with every other ID prop in the database; this makes chat messages
consistent.