Record guess state transitions in the chat log for visibility
It's awkward for people to watch in the answer submission dialog for possibly
multiple minutes to see if their answer was correct or not.  Having a message
appear in chat is a better experience that lets you return to the puzzle and
double-check your work and still get notified immediately if you were
right/wrong instead of sitting idle while you wait for HQ to call in your
guess.