Refactor how we track types of documents
We effectively want two layers of typing for documents - a high-level
"provider" (Google Docs being one; our internal implementation of
spreadsheets being another if we ever introduce it), and within the
Google provider, a more specific type (sheets or docs). The
expectation is that within a provider, a lot of logic is shared.

This seems a little like it's overfitting, but it should make
introducing doc support much easier.