avoid componentWillMount in NavAggregator
This was a bit more involved -- we can't rely on `componentDidMount` to get
called in DOM-depth order, nor reverse DOM-depth order, so we handwave it in
situations that the behavior appears to work and provide an escape hatch to
provide an explicit sort key on breadcrumbs otherwise.

We could potentially make depth required and just provide it everywhere, which
probably wouldn't be the end of the world.  I think my main goal with the
navAggregator hacks were to make it so we populated the data in the navbar in
one place, and that place was under the component that subscribed to data and
thus had e.g. hunt/puzzle names available.