Avoid unsafe lifecycle methods in PuzzlePage/ChatHistory
We don't actually need to call shouldSaveScroll hook on componentWillUpdate;
it's already set in advance.  If we later discover my testing was inadequate
and we need this after all, we can use `getSnapshotBeforeUpdate()` per
https://reactjs.org/docs/react-component.html#getsnapshotbeforeupdate

We can move the ensureDocumentAndPermissions call into componentDidMount.

We can also move the other one (on navigation between puzzle pages) into
componentDidUpdate().