Avoid componentWillReceiveProps in PuzzleModalForm
We add "dirty" logic to determine if we should use the latest props
or if we should be rendering the data stored in this.state.

This actually introduces a small bug -- because ReactSelect2 emits the
`onChange` event more frequently than the user actually changes the values of
the form (in particular, it emits the event once on first render) the tags
field is marked dirty, regardless of if the user has edited them or not.

This means that if you have shown the form for editing a puzzle and hidden it,
or even edited and submitted it, if you open the form again, it may have stale
data.  This seems tolerable, especially since most of the time people seem more
likely to use the inline tag editor for existing puzzles.

I think this issue will likely go away when we replace ReactSelect2 with a less
jQuery and more React implementation, which I hope to get to once we've
upgraded React.