Avoid componentWillReceiveProps in SplitPanePlus
There was a feature involving a prop `size` that we never appear
to use.  Additionally, tracking `collapsed` both internally and
from the parent as props seems like a rats nest; if you want things
to collapse, save the state about it outside of `SplitPanePlus`.
(Which we already do everywhere we use the collapse feature.)

By eliminating these behaviors, the componentWillReceiveProps function
becomes a noop, and we can safely eliminate it entirely.