Try adding `<meta>` tags to `<head>` dynamically
We appear to have lost the mobile behavior that restricts width to the actual
device width.  I suspect this is because we lost the meta tags that were once
from head.html.

Try adding them dynamically with JS and see what browsers do.  It's not obvious
to me if under Blaze this would have been compiled into the initial HTML
fetched or if it would have dynamically added it like I'm doing here.