Fix glyphicons...again
This time, stop trying to import them, and instead just put them into
client/stylesheets/. This causes them to be compiled into the merged
stylesheet, instead of added dynamically via Javascript. In theory,
either should work, because we specified an absolute path for fonts.
However, in practice, it seems that the Meteor compiler is stripping the
leading slash (making the paths relative) on production builds.

This is annoying, but it's acceptable for the compiled stylesheet, since
that's served out of the root of the domain, meaning that the font files
are correctly positioned relative to it.