Upgrade to pre-release babel-eslint
babel-eslint v11 supports setting custom configuration for Babel, which
is valuable, because it lets us enable Babel's native Typescript
support. This seems to be the future of Typescript linting (Microsoft
has committed to abandoning TSlint in favor of ESlint).

v11 does require us to ship our own babelrc, though. This is designed to
mirror what the previous config was doing.