Show worksheet link on tablets
Even in "desktop" (wide) mode, show the worksheet link on tablets because Google Sheets no longer supports in browser edit mode.

Showing the link based on any-pointer: fine media query.  This is likely not precisely the criteria used by google to decide if the sheet can be edited in browser or not, but it works for iPads, which is the dominant area of concern.  Could simply always show the link, but @ebroder expressed a concern that doing so could encourage players to work outside of Jolly Roger.

Still always shows the worksheet link when in narrow mode.