Fix lint
The worksheet link really does behave as a link, but js first attempts to open in the native app if possible.  This creates a problem for anchor-is-valid.  To satisfy that, give the link the actual target, then cancel in js.  Open to suggestions if there's a better way to handle this.

Also fix import ordering.