Suppress sender in chat if same as previous message and sent in same minute
This can save screen real estate by avoiding showing tons and tons of lines
with display names and timestamps when they're the same person talking a lot.
I want to show sender + timestamp if the senders differ (for correctness) or if
the timestamps differ by more than, say, 60 seconds.  This strategy is not
particularly sophisticated; I could imagine fancier rules for when we suppress
showing senders like tracking the time of the last-timestamp-shown message and
seeing if that has gotten large enough, but that's unlikely to matter a lot
unless people primarily talk entirely to themselves in chat.

This could get a little computationally expensive in the event we have to
reparse many dates for puzzles with very long chat histories where the same
person writes a lot of messages in a row.  I tried to order the guard clauses
to make the comparisons as cheap as possible; I don't think it'll be a problem
in practice, but if we notice spikes in client-side latency from chat messages,
consider finding a way to optimize this change or just revert this commit.