Try to fix a render failure
When adding new puzzles, we often create new tags at the same time.  The Puzzle
can arrive over DDP before the Tag, so there's a window where we might render
with a tag ID for which we don't have a corresponding Tag object.  Annoyingly,
this makes React drop all future rendering and the user has to refresh the page
to recover.

Attempt to work around this by guarding the dereference on the tag actually
existing.