Loosen the types on collection.insert
Because we're using SimpleSchema autoValues in a lot of cases, those
fields aren't actually required for inserts. But the type specification
for insert requires a full object, and there's no way to communicate
through the types that some types are auto-populated (i.e. required to
be present on read, but not on write).

Work around this by overriding the type for insert to be significantly
looser. This is actually loose to the point of being technically unsafe,
but there's no way in the type layer to tell which fields have autoValue
specified and which don't.