Convert most migrations to TS
There's a little shenanigans here, because a bunch of the queries or
updates reference versions of the database schema that aren't current. I
decided to handle this by just casting a bunch of things to `any`. In
theory, we could maybe do something smarter with some notion of
versioned schema types, but that seems challenging to manage. This is
probably fine, all things considered.

Alternatively, maybe we should just delete migrations that reference
invalid schemas.