Add support for Nitro instances
Nitro exposes block devices via NVMe, and cloud-init doesn't know how to
address block devices in a way that's consistent across Nitro and Xen
instances. So we have to do it ourselves.

Ship a script that can take an EBS device name as supplied to the EC2
API (like "sdf") and translate it into an actual block device that
exists. Use that to set up swap.