Update @types/meteor
This makes some of the types more accurate, especially `Meteor.user()`,
`Meteor.userId()`, and `Collection.findOne()`, all of which can return
null, but previously weren't typed to do so.

For user and userId, in most cases we know whether or not we're logged
in based on the current active component, so this can mostly be handled
by the type assertion operator. Similarly, there are several cases where
we don't need to handle null if we assume our database is consistent.
However, some of these changes did catch actual areas where we didn't
have strong enough null guards.