Fix rendering of breadcrumbs, broken in #256
LinkContainer decides whether or not a given link is active, and then
forces a value for the active prop onto its child (ugh). By default, it
considers itself "active" even when its on a child page, but
onlyActiveOnIndex overrides that behavior, ensuring that it only
considers itself active unless we're actually on /hunts/:hunt (which we
never are - we always redirect to puzzles, but it's easy to go from
active=false to active=true and hard to go the other way).

(Also fix the styling, which was getting mis-applied to the wrong
elements)