Lighten announcement and guess subscriptions
Historically, we've made overly broad subscriptions to have information
ambiently available to us. But that results in fairly heavy data fetches
on initial page load (e.g. we fetch all puzzles just in case there's a
pending guess).

By pushing some of the logic to the server, and having a more specific
subscription, we can significantly tighten the scope of data that we
have to fetch.