Increase chat textarea lineHeight
The short lineheight in the textarea did not match chat styling directly above and hurt legibility. Note that fontSize in the textarea is higher than in chat history in order to prevent automatic zooming in mobile Safari (see #97). This change matches the lightHeight to the chat history in absolute, not relative terms, which I found to be the more aesthetic option.