Allow answers in puzzle metadata to break
Previously, answers in puzzle metadata did not break, which could cause very narrow breaking of titles for puzzles with long answers, especially on mobile or other narrow windows. Simply allowing the answer to break (combined with equal flex grow on both parts) solves the problem in a reasonable - albeit not optimal - way, without requiring onResize logic. This behavior is fairly consistant with the existing behavior on the puzzle list.