Remove the Meteor subs-cache
This component seems to be a little bit flaky, and in most cases we're
using the same subscriptions across many different pages. This does
require tweaking App to be a single component, instead of swapping
between the old FullscreenLayout and ScrollableLayout components, since
otherwise any time we change layout looks like a full DOM replacement.

I also made one small change in subscription logic in HuntApp to make
sure that one of the subscriptions was reusable for sub-components.