Automatically collapse related puzzles if unlikely to be useful
Per Issue #234, related puzzles are most useful for meta puzzles. For other puzzles, the information should be available, but the space may be better devoted to chat. This automatically collapses related puzzles for puzzles without meta-related tags. When navigating between puzzles, related puzzles will never be collapsed automatically, but may be expanded (though currently there are no links to other puzzles other than in the related pane, so the latter is unlikely to occur).