Relocate viewers in puzzle metadata
Despite doing the same thing (opening a modal), the viewers and guess button have different presentations. The former is an unstyled link with a popover and the latter is a button. Since the action isn't navigation, standardize on button. Also move the viewers button the bottom row, as it doesn't make a lot of sense inline with the title and contributes to needless line breaks there.

Change the  text for both buttons to be more concise (and invoke a verb). The viewers button now reads "See <n> hunter(s)" while the guess button reads either "Guess (<n> so far)" or "See <n> guess(es)" depending on the solve status.

Change the labels for document links to reflect Google's naming convention (which conveniently shortens them).

With those changes, everything fits on one line on most modern phones (using iPhone 8 as a guide). Below 375px (e.g. iPhone SE at 320px), the link labels are hidden. The link labels are hidden with css media queries, which is sufficient as long as the desktop view isn't allowed to reduce the width below 375px.