Get rid of desiredLayout hack
If we're a bit more clever with our CSS (and accept one very minor UI
change - namely that the disconnect alert is always at a fixed
non-scrolling position), then we can have fixed markup for the app
framework and only need custom markup and styling for the actual puzzle
page.

This gets rid of the fairly awkward hack where the App component was
going digging through the resolved routes under it (which will make the
react-router v5 upgrade easier).