Fix autovalue stuff in puzzles around arrays
autoValue is confusing and sometimes gets the whole array (on document
creation?) and sometimes gets just the element that changed (a string, in this
case).  Loosen the types and do some runtime type-checking to deal with it.

Fix a typo in guesses where we wouldn't pull the guess from `answers` when
transitioning from correct to any other state..