Tweak search on PuzzleListPage
I'm not sure why this was disjunctive before, but it doesn't make a ton of
sense to me now.  Now, if you provide multiple space-separated search keys, the
filter will show only puzzles that match all of the keys in some way, which I
think aligns better with "typing is specifying additional constraints over what
we intend to show", and means that typing more should never result in seeing a
larger result set.

Additionally, arbitrary substring search in puzzle title might be a bit much.
We can preseve the above property and narrow in more quickly if we do word-prefix
search instead.  We'll use this for puzzle titles, but answers and tags are often
written with no whitespace, so we'll stick with arbitrary substring matching for
those.