webrtc demo: add analyserNode & spectrogram
This does an FFT on the audio from each peer.

The intended purpose of this computation is to allow us to automatically detect
when a peer is speaking, so we can mark that peer in the user list as actively
speaking.

The other feature we can potentially use this for is to potentially squelch
feedback between peers.  If the dB level is too high, we can potentially
automatically adjust gain in the gainNode to protect people's ears.