Invite users to configured guild when Discord account linked
This implements Discord OAuth2 atop Meteor's reusable OAuth tools.

We add a button on the user's profile page to allow them to link their Discord
account to Jolly Roger.  This initiates an OAuth flow to request a token with
the `identify` and `guilds.join` permissions.  Upon successful completion of
the OAuth flow, we persist the OAuth token on the User object under `services`,
look up the user's ID via the Discord API, and save the user's handle and
discriminator on their Profile object.

Then, additionally, we check if the user is a member of the server-wide
configured guild.  If they are not, we add them to the guild via the bot
account.