Add spectrograms, muted/deafened icons
We could probably make the icons black and draw a red diagonal line through
them but this is the limit of my creativity for the night.

We could stand to either feature-flag the spectrograms, make them
user-toggleable, or just disable them after a certain number of users, since
they seem to take up a decent bit of CPU time to render.