Make deafened imply muted
It seems like a bad thing to allow someone to be generating noise for others
but not be able to see that they are doing so.  To the extent that we can, we
should avoid admitting this state.  So let's make this effectively a tristate:
full active call participation, muted, or deafened.  (I've kept it split out as
booleans in the datamodel for simplicity of rendering.)

For what it's worth, Discord appears to behave similarly, which is another
point in the "someone thought about this and decided it was the right way to
design this" column.