Show all tabs of rtcCallers
This prevents the "I joined and a half dozen additional people showed up"
surprising behavior.

There's still one thing that will look weird, which is that if you have two
tabs open to the same puzzle with the same user and one is in the call and the
other is not, you'll see yourself in the call participants list but not in
the viewers list, despite the tab you're viewing things from is just a viewer.

This is fine, in my opinion.  The presence tracking is primarily intended to be
useful for *other people*, and this arrangement will do entirely reasonable
things when viewed by any other user.