Add in-app first-user-creation flow for server admin
It's actually kinda hard to set up a new production Jolly Roger deployment,
since there's no good way to get a meteor shell with a connection to the real
DB to create the first account.  You either have to figure out how to get a
shell hooked up to your production DB (which we don't document), or you have to
bring a row from your users table via a mongo client, or something else
ridiculous.  We can do better.

This change adds:

* a pseudocollection that simply indicates whether there are any
  users in the Meteor users table
* a Meteor method to provision the initial admin user provided there
  are no users in the Meteor users table yet
* a page with a form to enter an email address/password to be used with the
  aforementioned Meteor method
* logic in the root redirector to send you to that form if you are not logged in
  and the server has no users

Then, when setting up a new server, you navigate to the root page, it redirects
to the first-account-creation page, and you can just create the admin user
through the UI.