Add missing `server` field to `CallParticipants` schema
This was unfortunately subtle and undetected by any of our usual safeguards.

The field was provided to CallParticipants.insert, but because it was not
declared in the schema, it was ignored and never serialized to the database.
This broke CallParticipant garbage collection, as no records were tagged with
the server they came from, so the cleanup by dead-server logic would never
detect these records as orphaned.

Somewhat surprisingly, neither our TypeScript tools nor Meteor's runtime code
noticed that we were passing additional key/values not present in the schema to
complain, but instead silently stripped them out.

Ideally we wouldn't need to expose the server field to the client, but I'm not
sure how to phrase that to the type system.