Update acceptance tests
I hate adding this sleep rather than some sort of thing that we observe until
success (with a timeout), but I'm not sure what else we'd be able to observe
or if Meteor gives us better tools for this sort of thing.

Adding `--full-app` to the `meteor test` invocation seemed to be necessary to
get the subscriptions to work, at which point it occurred to me that we
probably ought to be rerunning migrations after we reset the database.