Support TURN for peer discovery
This change shifts how RTC configuration is done, enabling us to use
authenticated TURN servers (which unlocks the ability to route traffic for
peers that cannot reach each other directly).

To accomplish this, the TURN server needs to be able to authenticate its
clients.  While there's a bunch of complicated ways to do this, there's also
one pleasantly simple approach in [1] that requires no more shared state than a
single shared secret between a web service (for us, the Meteor backend) and the
TURN server (for us, `coturn`).

We implement the token generation scheme with HMAC-SHA1 as specified, and
publish the resulting username/credential in a pseudocollection called
`rtcconfig`.  Since now TURN integration requires a secret, and since we're not
giving TURN config to non-authenticated users, it's simpler to drop the whole
`PublicConfig` thing that I wrote for the stun server URLs and just stuff the
whole thing in Settings on the server and in the pseudocollection to get it to
the client.  Fewer, fatter structs with shared lifecycle.

Server administrators will need to update their configuration to include the
TURN shared secret and to specify STUN/TURN urls appropriately.

While deployment guides suggest shorter-lived tokens, I'm not worried about
abuse from our users so I made the tokens valid for 3 days, which massively
exceeds the maximum time I imagine any one tab will have joined a call.

You can verify that relaying actually works with your setup by adding
`iceTransportPolicy:"relay"` to the `rtcConfig` passed to `new
RTCPeerConnection` in `CallLinkBox.tsx`, which will force the
`RTCPeerConnection` to only use relay candidates, and then verifying that calls
still work.

[1] - https://tools.ietf.org/html/draft-uberti-behave-turn-rest-00